PTCD-866 Remove legacy IVenueService (#1976)
* Remove legacy IVenueService

* Replace legacy Venue model with Core Venue model

+++ src/Dfc.CourseDirectory.Core/DataStore/CosmosDb/QueryHandlers/DeleteVenueHandler.cs
@@ -34,7 +34,7 @@ namespace Dfc.CourseDirectory.Core.DataStore.CosmosDb.QueryHandlers
 
                 venue = query.Document;
 
-                venue.Status = (int)VenueStatus.Archived;
+                venue.Status = VenueStatus.Archived;
                 venue.DateUpdated = request.UpdatedDate;
                 venue.UpdatedBy = request.UpdatedBy.UserId;
